made instrumentation more DRY
